reorganized Code base
The focus for the application part of Gdv should be a very tiny viewer that simply uses
a File-descriptor and displays all data that can be extracted from that file-descriptor.

The idea to create a full data graphing and analysis application will be shifted to a
different project with the name "rhodium".